+++ packages/core/src/__tests__/data/for/basic-for-fragment.raw.tsx
@@ -1,6 +1,10 @@
-import { For, Fragment } from '@builder.io/mitosis';
+import { For, Fragment, useStore } from '@builder.io/mitosis';
 
 export default function BasicForFragment() {
+  const state = useStore({
+    id: 'xyz',
+  });
+
   return (
     <div>
       <For each={['a', 'b', 'c']}>
@@ -10,6 +14,22 @@ export default function BasicForFragment() {
           </Fragment>
         )}
       </For>
+      <For each={['a', 'b', 'c']}>
+        {(option) => (
+          <Fragment key={`${state.id}-${option}`}>
+            <div>{option}</div>
+          </Fragment>
+        )}
+      </For>
+      <select>
+        <For each={['d', 'e', 'f']}>
+          {(option) => (
+            <option key={`${state.id}-${option}`} value={option}>
+              {option}
+            </option>
+          )}
+        </For>
+      </select>
     </div>
   );
 }
